jobsdb/requirements.txt
asgiref==3.2.3
Django==2.2.8
django-ckeditor==5.8.0
django-js-asset==1.2.2
pytz==2019.3
sqlparse==0.3.0
